# openapi-crm-springboot.yaml
openapi: 3.0.3
info:
  title: CRM Backend API
  version: 1.0.0
servers:
  - url: /api/v1
paths:
  /auth/login:
    post:
      tags: [Auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
  /auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
...
  /contacts:
    get:
      tags: [Contacts]
      summary: List contacts
      responses:
        '200':
          description: Contacts
    post:
      tags: [Contacts]
      summary: Create contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactRequest'
      responses:
        '201':
          description: Contact created
  /contacts/{id}:
    get:
      tags: [Contacts]
      summary: Get contact by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Contact
    put:
      tags: [Contacts]
      summary: Update contact
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactRequest'
      responses:
        '200':
          description: Contact updated
    delete:
      tags: [Contacts]
      summary: Delete contact
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Contact deleted
...
  /teams/{id}:
    get:
      tags: [Teams]
      summary: Get team by id
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Team
    put:
      tags: [Teams]
      summary: Update team
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                teamName:
                  type: string
      responses:
        '200':
          description: Team updated
    delete:
      tags: [Teams]
      summary: Delete team
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Team deleted
...
components:
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    LoginRequest:
      type: object
      required: [usernameOrEmail, password]
      properties:
        usernameOrEmail:
          type: string
        password:
          type: string
    ChangePasswordRequest:
      type: object
      required: [oldPassword, newPassword, confirmPassword]
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
        confirmPassword:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          format: int64
        user:
          type: object
          additionalProperties: true
        permissions:
          type: array
          items:
            type: string
...
    ContactRequest:
      type: object
      properties:
        contactName: { type: string }
        contactMobileNo: { type: string }
        contactEmail: { type: string }
        contactAddress: { type: string }
        contactCity: { type: string }
        contactState: { type: string }
        contactCountry: { type: string }
        contactOccasion: { type: string }
        contactPostalCode: { type: string }
        contactOccasionDate: { type: string, format: date }
...